fix viral import method
Adding our package to @ISA adds not just the intended
MODIFY_CODE_ATTRIBUTES method, but also any other methods in the
package, including import. This would result in the importing package
gaining its own import that would spread the :Pure attributes.

Fix this by using a separate package for the MODIFIY_CODE_ATTRIBUTES
method, and adding it to @ISA instead. Use a fully qualified sub
definition so that it stays in Attribute::Pure context and can call its
subs.