generate a Build.PL in the repo root for ease of testing
Create a Build.PL in the repository root so testing can be done without
needing to invoke Dist::Zilla. Regenerate this using [@Starter]'s
regenerate option, to update the file on release or via the dzil
regenerate command.

Add an additional entry to the gitignore for the temporary .c files
created if this Build.PL is used.